requirements.txt
streamlit==1.35.0
anthropic==0.97.0
python-dotenv==1.0.0
reportlab==4.1.0